# Smarter error collecting