Only populate the `expected` HashSet when we know that parsing will
fail. This results in impressive speedups in the non-failing case:

```
Before:
test expr ... bench:      5926 ns/iter (+/- 621) = 4 MB/s
test json ... bench:     24909 ns/iter (+/- 575) = 10 MB/s

After:
test expr ... bench:      4415 ns/iter (+/- 220) = 5 MB/s
test json ... bench:      6565 ns/iter (+/- 526) = 41 MB/s
```

Note that this will make parsing slower if it fails, since it basically parses the input twice (both parser invocations individually are faster, though). I don't think this is an issue, since when a program fails to parse something, it will usually report an error and abort (ie. its a slow path anyway).

I also experimented with replacing the default `SipHasher` used for HashMaps and HashSets with the faster Murmurhash2, which brought the `expr` benchmark up to 16 MB/s (while completely commenting out `expected` handling, so actual gains will be a bit lower). This has a few tradeoffs, since hasher replacement is still experimental in Rust, so I don't know if I'll go in that direction any further.